Ejercicio 1 Resuelto y funcional
